remove unnecessary model concerns, move validation and feat on their respective model
