--- street-goods-burger-backend/app/models/concerns/store_transaction_validations.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-# StoreTransactionValidations module
-module StoreTransactionValidations
-  extend ActiveSupport::Concern
-  include ActiveModel::Validations
-
-  included do
-    belongs_to :store_customer
-    has_one :cart
-
-    validates :status,
-              presence: true,
-              inclusion: {
-                in: %w[pending pre_process processing delivered canceled],
-                message: 'Invalid status value'
-              }
-  end
-end
