+++ street-goods-burger-backend/app/models/food_order.rb
@@ -2,5 +2,10 @@
 
 # FoodOrder template
 class FoodOrder < ApplicationRecord
-  include FoodOrderValidations
+  belongs_to :cart
+
+  validates :food_name, length: { minimum: 1, maximum: 64 }
+  validates :food_category, presence: true, length: { minimum: 1, maximum: 64 }
+  validates :price, presence: true, numericality: { greater_than: -0.99 }
+  validates :quantity, presence: true, numericality: { only_integer: true, greater_than: 0 }
 end
