+++ street-goods-burger-backend/app/models/review.rb
@@ -2,5 +2,8 @@
 
 # Review Template
 class Review < ApplicationRecord
-  include ReviewValidations
+  belongs_to :food
+  belongs_to :store_customer
+
+  validates :text_body, length: { minimum: 1, maximum: 256 }
 end
