+++ street-goods-burger-backend/app/models/store_transaction.rb
@@ -2,5 +2,13 @@
 
 # StoreTransaction Template
 class StoreTransaction < ApplicationRecord
-  include StoreTransactionValidations
+  belongs_to :store_customer
+  has_one :cart
+
+  validates :status,
+            presence: true,
+            inclusion: {
+              in: %w[pending pre_process processing delivered canceled],
+              message: 'Invalid status value'
+            }
 end
